fix specs
